% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_forecast.R
\name{calib_xgb}
\alias{calib_xgb}
\title{用 XAJ calib 历史模拟结果重新率定 XGB 后处理模型}
\usage{
calib_xgb(
  site = NULL,
  f_xgb = NULL,
  f_calib = NULL,
  nlead = 12,
  force_calib = FALSE
)
}
\description{
准备 XGBoost forecast 后处理模型
force_calib = TRUE 表示用 XAJ calib 文件重新率定；FALSE 表示读取已有 res。
}
