% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroFloods-package.R
\docType{package}
\name{HydroFloods-package}
\alias{HydroFloods}
\alias{HydroFloods-package}
\title{HydroFloods: Tools for hydrological model}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
