% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Forecast.R
\name{plot_Forecast}
\alias{plot_Forecast}
\title{预报窗口图：过去一周 + 未来n小时，凸显t0位置与预报时段}
\usage{
plot_Forecast(
  t0,
  fout_sim,
  fout_xgb = NULL,
  prcp.max = NULL,
  q.max = NULL,
  window_past = days(7),
  window_fc = days(1)
)
}
\arguments{
\item{t0}{预报起点，POSIXct 或可被 as.POSIXct 解析的字符串}

\item{fout_sim}{预报阶段，X2对应的全部模拟结果, 必须包含 time, Q_obs, Q_sim, P列}

\item{fout_xgb}{可选，XGBoost后处理的结果文件，包含 time,
XGB_mean, XGBQlag_mean, XGBQlagMulti_mean}

\item{window_past}{回溯时长（默认7天）}

\item{window_fc}{预报时长（默认1天）}
}
\description{
预报窗口图：过去一周 + 未来n小时，凸显t0位置与预报时段
}
