% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_forecast.R
\name{run_xgb_forecast}
\alias{run_xgb_forecast}
\title{XAJ window 的 XGBoost 后处理主函数
xaj_window 必须已包含 PET 或 PET_Romanenko。}
\usage{
run_xgb_forecast(
  site = NULL,
  f_fcWin,
  f_xgb = NULL,
  f_calib = NULL,
  fout = "OUTPUT/XAJ_XGB_forecast.csv",
  nlead = 12,
  force_calib = FALSE
)
}
\description{
XAJ window 的 XGBoost 后处理主函数
xaj_window 必须已包含 PET 或 PET_Romanenko。
}
