% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_parameter_score.R
\name{train_xgboost_par}
\alias{train_xgboost_par}
\title{单站点 XGBoost 订正模型训练}
\usage{
train_xgboost_par(
  data_full,
  leads = 1:12,
  nrounds = 200,
  early_stopping_rounds = 20,
  max_depth = 3,
  min_child_weight = 5,
  eta = NULL,
  subsample = 1,
  colsample_bytree = NULL,
  lambda = NULL,
  gamma = NULL
)
}
\description{
单站点 XGBoost 订正模型训练
}
