% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_validate.R
\name{eval_floods}
\alias{eval_floods}
\title{洪水场次合格率 (cal_pass_rate / summary_xgboost 共用)}
\usage{
eval_floods(pred, d_full)
}
\arguments{
\item{pred}{含 \verb{time, model, lead, kfold, Q_obs, Q_sim} 列的长表;}

\item{d_full}{划分洪水场次的全序列}
}
\description{
洪水场次合格率 (cal_pass_rate / summary_xgboost 共用)
}
