% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FloodEvents_Qsim_models.R
\name{plot_FloodEvents_Qsim_models}
\alias{plot_FloodEvents_Qsim_models}
\title{plot_FloodEvents_Qsim_models}
\usage{
plot_FloodEvents_Qsim_models(
  dat_obs,
  dat_sim,
  prcp.max = 50,
  q.max = 250,
  ...,
  colors = NULL,
  ncol = 4,
  show_passed = TRUE
)
}
\arguments{
\item{dat_obs}{观测的时间序列数据，必须包含 \code{time} 和 \code{Q_obs}, \code{group_name}}

\item{dat_sim}{模拟的时间序列数据，必须包含 \code{time}、\code{Q_sim}, \code{Q_obs}和 \code{model} 三列}
}
\description{
plot_FloodEvents_Qsim_models
}
